tests/wpt/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-allowfullscreen.html

<!doctype html>
<meta charset=utf-8>
<title>Check how allowfullscreen affects fullscreen enabled flag</title>
<link rel="author" title="Xidorn Quan" href="https://www.upsuper.org">
<link rel="author" title="Mozilla" href="https://www.mozilla.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsers.html#initialise-the-document-object">
<link rel="help" href="https://fullscreen.spec.whatwg.org/#fullscreen-enabled-flag">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>
<script>
  async_test(function(t) {
    var iframe = document.createElement("iframe");
    iframe.src = "support/blank.htm";
    var eventWatcher = new EventWatcher(t, iframe, "load");
    document.body.appendChild(iframe);
    t.add_cleanup(function() {
      document.body.removeChild(iframe);
    });

    assert_true(document.fullscreenEnabled, "Top level document has fullscreen enabled flag set");
    eventWatcher.wait_for("load").then(t.step_func_done(function() {
      assert_false(iframe.contentDocument.fullscreenEnabled, "Document inside iframe without allowfullscreen attribute should not have fullscreen enabled flag set");
      iframe.setAttribute("allowfullscreen", true);
      assert_true(iframe.contentDocument.fullscreenEnabled, "Fullscreen should be allowed when allowfullscreen attribute is set");
      iframe.removeAttribute("allowfullscreen");
      assert_false(iframe.contentDocument.fullscreenEnabled, "Fullscreen should be denied when allowfullscreen attribute is removed");
    }));
  }, "iframe-allowfullscreen");

  /* Fullscreen enabled flag with about:blank */

  function test_allowfullscreen_noload(setup_iframe, check) {
    var iframe = document.createElement("iframe");
    setup_iframe(iframe);
    document.body.appendChild(iframe);
    check(iframe.contentDocument);
    document.body.removeChild(iframe);
  }

  test(function() {
    test_allowfullscreen_noload(function() {}, function(doc) {
      assert_false(doc.fullscreenEnabled, "Fullscreen should not be enabled without allowfullscreen attribute");
    });
  }, "iframe-noload-noallowfullscreen");

  test(function() {
    test_allowfullscreen_noload(function(iframe) {
      iframe.setAttribute("allowfullscreen", true);
    }, function(doc) {
      assert_true(doc.fullscreenEnabled, "Fullscreen should be enabled with allowfullscreen attribute");
    });
  }, "iframe-noload-allowfullscreen");
</script>
